+++ pages/api/v1/list/[listId]/index.ts
@@ -18,18 +18,37 @@ export default async function handle(req, res) {
   }
   if (req.method === "DELETE") {
     if(listId) {  
-      const list = await prisma.list.delete({
+      const list = await prisma.list.findUnique({
         where: {
           id: parseInt(listId, 10)
         }
       })
-      return res.json()
+      if(list.ownerEmail !== session.user?.email) {
+        console.log("You're not the owner")
+        return res.json({message: "You can only delete lists you own!"})
+      }
+      await prisma.listToUser.deleteMany({
+        where: {
+          listId: parseInt(listId, 10)
+        }
+      })
+      await prisma.list.delete({
+        where: {
+          id: parseInt(listId, 10)
+        }
+      })
+      return res.status(204).send()
     }
   }
   if (req.method === "PATCH") {
-    const { rename } = req.body
-    if(rename) {  
-      const list = await prisma.list.update({ 
+    const { rename, email, userToDelete } = req.body
+    const list = await prisma.list.findUnique({
+      where: {
+        id: parseInt(listId, 10)
+      }
+    })
+    if(rename && list.ownerEmail === session.user.email) {  
+      await prisma.list.update({ 
         where: {
           id: parseInt(listId, 10)
         },
@@ -38,6 +57,47 @@ export default async function handle(req, res) {
         }
       })
       return res.json()
+    }
+    if(email) {
+      try {
+        const userToAdd = await prisma.user.findUnique({
+          where: {
+            email: email
+          }
+        })
+
+        if(userToAdd) {
+          await prisma.list.update({
+            where: {
+              id: parseInt(listId, 10)
+            },
+            data: {
+              users: {
+                create: [
+                  { user: { connect: { email: email } } }
+                ]
+              }
+            }
+          })
+          return res.json({message: "User added successfully"})
+        } else {
+          return res.json({message: `User with email "${email}" does not exist`})
+        }
+      } catch(err) {
+        console.error(err)
+      }
+    }
+    if(userToDelete) {
+      const user = await prisma.user.findUnique({
+        where: { email: userToDelete }
+      })
+      await prisma.listToUser.deleteMany({
+        where: {
+          listId: parseInt(listId, 10),
+          userId: user.id
+        }
+      })
+      return res.status(204).send()
     }
   }
 }
